Convert to showing frames as attachment.
Switch from showing the frame as the message text and instead
display as an attachment. This gets around the 4K limit, and
is a little cleaner (we can show context around it, etc.).
Downside is you have to expand the attachment to watch (this
could also be seen as a benefit).